Bug fixed for large file.
git-svn-id: http://parabix.costar.sfu.ca/svn/icGREP/icgrep-devel@4973 a82c6ad6-7e1c-4be0-8e68-26c050b43eaf

+++ icgrep/grep_engine.h
@@ -15,7 +15,7 @@
 namespace llvm { class raw_ostream; }
 
 class GrepEngine {
-    typedef void (*GrepFunctionType)(char * byte_data, int filesize, const char* filename, uint64_t finalLineUnterminated);
+    typedef void (*GrepFunctionType)(char * byte_data, size_t filesize, const char* filename, uint64_t finalLineUnterminated);
 public:
 
     GrepEngine() {}
